fix(MinterGateway): return early if maxAllowedActiveOwedM is gt uint112

+++ src/MinterGateway.sol
@@ -433,7 +433,6 @@ contract MinterGateway is IMinterGateway, ContinuousIndexing, ERC712 {
     function excessOwedM() public view returns (uint240 excessOwedM_) {
         // NOTE: Can safely cast to `uint240` since we know M Token totalSupply constraints.
         uint240 totalMSupply_ = uint240(IMToken(mToken).totalSupply());
-
         uint240 totalOwedM_ = _getPresentAmountRoundedDown(principalOfTotalActiveOwedM, currentIndex()) +
             totalInactiveOwedM;
 
@@ -705,9 +704,9 @@ contract MinterGateway is IMinterGateway, ContinuousIndexing, ERC712 {
 
         uint256 maxAllowedActiveOwedM_ = maxAllowedActiveOwedMOf(minter_);
 
-        // If the minter's max allowed active owed M is greater than `type(uint240).max`, then it's definitely greater
-        // than the max possible active owed M for the minter, which is capped at `type(uint240).max`.
-        if (maxAllowedActiveOwedM_ >= type(uint240).max) return;
+        // If the minter's `maxAllowedActiveOwedM_` is greater or equal to `type(uint112).max`,
+        // we return early since it would revert when computing the `principalOfMaxAllowedActiveOwedM_`.
+        if (maxAllowedActiveOwedM_ >= type(uint112).max) return;
 
         // NOTE: Round the principal down in favor of the protocol since this is a max applied to the minter.
         uint112 principalOfMaxAllowedActiveOwedM_ = _getPrincipalAmountRoundedDown(uint240(maxAllowedActiveOwedM_));
